% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{Assess.covariate.importance}
\alias{Assess.covariate.importance}
\title{Compute correlation between parameters and an outcome variable}
\usage{
Assess.covariate.importance(don, X, Y, method)
}
\description{
Compute correlation between parameters and an outcome variable
}
