% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{assess.parameter.importance}
\alias{assess.parameter.importance}
\title{Assess parameter importance}
\usage{
assess.parameter.importance(don, X, Y, method)
}
\arguments{
\item{don}{a data frame.}

\item{X}{a numerical vector.}

\item{Y}{a numerical vector.}

\item{method}{A string}
}
\value{
a vector.
}
\description{
Assess parameter importance
}
