% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_functions.R
\name{ever.been.here}
\alias{ever.been.here}
\title{Another SEIR model function}
\usage{
ever.been.here(SEIR.object)
}
\arguments{
\item{SEIR.object}{A SEIR model object}
}
\description{
Another SEIR model function
}
