% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fetch.package}
\alias{fetch.package}
\title{Fetch a package}
\usage{
fetch.package(
  nom_du_package,
  try.this = c("nothing", "load", "install"),
  verbose = FALSE
)
}
\arguments{
\item{nom_du_package}{A string}

\item{try.this}{A string}

\item{verbose}{A logical}
}
\description{
Fetch a package
}
