% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesChunks.R
\name{from.tbl.to.df}
\alias{from.tbl.to.df}
\title{Convert a tibble to a data frame}
\usage{
from.tbl.to.df(data.frame.ish, rename = NULL)
}
\arguments{
\item{data.frame.ish}{A tibble}

\item{rename}{A vector or data frame}
}
\description{
Convert a tibble to a data frame
}
