% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.functions.R
\name{get.path}
\alias{get.path}
\title{Return the complete path to a demo model Excel workbook}
\usage{
get.path(x)
}
\arguments{
\item{x}{a character element representing a file name ("demo.1.xls", "demo.3.xls", "demo.4.xls").}
}
\value{
a character element representing the complete path to the file. If the file cannot be found, an empty element will be returned.
}
\description{
Return the complete path to a demo model Excel workbook
}
\examples{
path.to.demo.model <- get.path("demo.1.xls")

}
