% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\name{get.scatter.plot}
\alias{get.scatter.plot}
\title{Render a scatter plot}
\usage{
get.scatter.plot(
  x,
  y,
  x_label_text = deparse(substitute(x)),
  y_label_text = deparse(substitute(y)),
  geom_point_size = 2,
  element_text_size = 12,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{x_label_text}{a character string.}

\item{y_label_text}{a character string.}

\item{geom_point_size}{a numeric value.}

\item{element_text_size}{a numeric value.}

\item{width}{an integer.}

\item{height}{an integer.}
}
\value{
none.
}
\description{
Render a scatter plot
}
\examples{
# Load demo data
data("sweepr.demo")

# Define results
outcomes.summary.df <- sweepr.demo$results

get.scatter.plot(
  x = outcomes.summary.df$delta.overwrite, 
  y = outcomes.summary.df$maxInc, 
  height = 500, 
  width = 756
)

}
