% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{get.silly.model.chunk}
\alias{get.silly.model.chunk}
\title{Get SEIR model inputs}
\usage{
get.silly.model.chunk(
  model_flows,
  init.cond,
  init.cond.numeric.vars,
  which.flow,
  NewFrom,
  NewTo
)
}
\description{
Get SEIR model inputs
}
