% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\name{get.tornado.plot}
\alias{get.tornado.plot}
\title{Render a tornado plot}
\usage{
get.tornado.plot(
  outcome_variable,
  parameters = parms.tried.df,
  outcomes = outcomes.summary.df,
  method = "kendall-partial-correlation-slow",
  bin_width = 0.5,
  element_text_size = 12,
  order_by_absolute_value = FALSE,
  add_label = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{outcome_variable}{a character string.}

\item{parameters}{the parms.tried.df data frame.}

\item{outcomes}{the outcomes.summary.df data frame.}

\item{method}{a character string. Options include: kendall-partial-correlation-slow, pearson-partial-correlation-fast, pearson-partial-correlation-slow, spearman-partial-correlation-slow, t-test. By default, the option is kendall-partial-correlation-slow.}

\item{bin_width}{a numeric value.}

\item{element_text_size}{a numeric value.}

\item{order_by_absolute_value}{a logical value.}

\item{add_label}{a logical value.}

\item{width}{an integer.}

\item{height}{an integer.}
}
\value{
none.
}
\description{
Render a tornado plot
}
\examples{
# Load demo data
data("sweepr.demo")

# Define results
parameters.swept <- sweepr.demo$parameters
outcomes.summary.df <- sweepr.demo$results

get.tornado.plot(
  outcome_variable = "maxInc", 
  parameters = parameters.swept, 
  outcomes = outcomes.summary.df, 
  height = 500, 
  width = 756
)

}
