% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.functions.R
\name{get.tornado.table}
\alias{get.tornado.table}
\title{Render a tornado table}
\usage{
get.tornado.table(
  outcome_variable,
  parameters = parms.tried.df,
  outcomes = outcomes.summary.df,
  method = "kendall-partial-correlation-slow"
)
}
\arguments{
\item{outcome_variable}{a character string.}

\item{parameters}{the parms.tried.df data frame.}

\item{outcomes}{the outcomes.summary.df data frame.}

\item{method}{a character string. Options include: kendall-partial-correlation-slow, pearson-partial-correlation-fast, pearson-partial-correlation-slow, spearman-partial-correlation-slow, t-test. By default, the option is kendall-partial-correlation-slow.}
}
\value{
none.
}
\description{
Render a tornado table
}
\examples{
# Load demo data
data("sweepr.demo")

# Define results
parameters.swept <- sweepr.demo$parameters
outcomes.summary.df <- sweepr.demo$results

get.tornado.table(
  outcome_variable = "maxInc", 
  parameters = parameters.swept, 
  outcomes = outcomes.summary.df
)

}
