% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_scatter_plot}
\alias{get_scatter_plot}
\title{Render a scatter plot}
\usage{
get_scatter_plot(
  x,
  y,
  x_label_text = deparse(substitute(x)),
  y_label_text = deparse(substitute(y)),
  geom_point_size = 2,
  element_text_size = 12,
  width = 756,
  height = 500
)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{x_label_text}{A string}

\item{y_label_text}{A string}

\item{geom_point_size}{A numeric}

\item{element_text_size}{A numeric}

\item{width}{A numeric}

\item{height}{A numeric}
}
\description{
Render a scatter plot
}
