% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_tornado_plot}
\alias{get_tornado_plot}
\title{Render a tornado plot}
\usage{
get_tornado_plot(
  outcome_variable,
  parameters = parms.tried.df,
  outcomes = outcomes.summary.df,
  method = "kendall-partial-correlation-slow",
  bin_width = 0.5,
  element_text_size = 12,
  order_by_absolute_value = FALSE,
  add_label = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{outcome_variable}{A string}

\item{parameters}{The parms.tried.df data frame}

\item{outcomes}{The outcomes.summary.df data frame}

\item{method}{A string}

\item{bin_width}{A numeric}

\item{element_text_size}{A numeric}

\item{order_by_absolute_value}{A logical}

\item{add_label}{A logical}

\item{width}{A numeric}

\item{height}{A numeric}
}
\description{
Render a tornado plot
}
