% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_tornado_table}
\alias{get_tornado_table}
\title{Render a tornado table}
\usage{
get_tornado_table(
  outcome_variable,
  parameters = parms.tried.df,
  outcomes = outcomes.summary.df,
  method = "kendall-partial-correlation-slow"
)
}
\arguments{
\item{outcome_variable}{A string}

\item{parameters}{The parms.tried.df data frame}

\item{outcomes}{The outcomes.summary.df data frame}

\item{method}{A string}
}
\description{
Render a tornado table
}
