% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no.name.demo.R
\docType{data}
\name{no.name.demo}
\alias{no.name.demo}
\title{Box/compartment models results from 108 parameter sweeps.}
\format{
A list containing two data frames, outcomes.summary.df and parms.tried.df:
\describe{
\item{lambda.overwrite}{multiplier}
\item{delta.overweight}{multiplier}
...
}
}
\source{
\url{https://github.com/barnzilla/no.name}
}
\usage{
no.name.demo
}
\description{
A dataset containing box/compartment model results from 108 parameter sweeps.
}
\keyword{datasets}
