% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{overlap.length}
\alias{overlap.length}
\title{Overlap length function}
\usage{
overlap.length(L1, U1, L2, U2)
}
\arguments{
\item{L1}{A vector}

\item{U1}{A vector}

\item{L2}{A vector}

\item{U2}{A vector}
}
\description{
Overlap length function
}
