% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_functions.R
\name{SaveModelInExcel}
\alias{SaveModelInExcel}
\title{Save SEIR model as Excel file}
\usage{
SaveModelInExcel(input.info.list, file_name, map.names)
}
\arguments{
\item{input.info.list}{A data frame}

\item{file_name}{A string}

\item{map.names}{A vector string}
}
\description{
Save SEIR model as Excel file
}
