% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{SEIR.n.age.classes}
\alias{SEIR.n.age.classes}
\title{Main SEIR model functions}
\usage{
SEIR.n.age.classes(
  file.name,
  sheets.names,
  just.get.functions = FALSE,
  functions.kit = NULL,
  also.get.flows = NULL,
  agegrp.glue = "",
  CTMC.random.seeds = NULL
)
}
\arguments{
\item{file.name}{a string.}

\item{sheets.names}{a string vector.}

\item{just.get.functions}{a logical.}

\item{functions.kit}{a list containing one or more of the following: $differential.eqns.func, $post.processing.func, $post.processing.companion.kit, $CTMC.eqns.func, $CTMC.eqns.func.companion. This parameter is NULL by default.}

\item{also.get.flows}{a string.}

\item{agegrp.glue}{a string.}

\item{CTMC.random.seeds}{an integer vector.}
}
\value{
a SEIR object.
}
\description{
Main SEIR model functions
}
