% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{smooch.parms.df.into.list}
\alias{smooch.parms.df.into.list}
\title{Get other SEIR model inputs}
\usage{
smooch.parms.df.into.list(df.parms.1d, df.parms.2d, these.are.not.parms)
}
\description{
Get other SEIR model inputs
}
