% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{try.various.parms.values}
\alias{try.various.parms.values}
\title{Perform a parameter sweep}
\usage{
try.various.parms.values(
  SEIR.object,
  parm.cloud.grid.specs,
  only.show.parms.to.try = FALSE
)
}
\arguments{
\item{SEIR.object}{a SEIR model object.}

\item{parm.cloud.grid.specs}{a data frame of parameter sweep specifications.}

\item{only.show.parms.to.try}{a logical value. By default, FALSE.}
}
\value{
a list of parameter sweep inputs and results.
}
\description{
Perform a parameter sweep
}
\examples{
# Load demo data
data("sweepr.demo")

# Define results
outcomes.summary.df <- sweepr.demo$results

get.scatter.plot(
  x = outcomes.summary.df$delta.overwrite, 
  y = outcomes.summary.df$maxInc, 
  height = 500, 
  width = 756
)

}
