% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesChunks.R
\name{verbose.save}
\alias{verbose.save}
\title{Save R data}
\usage{
verbose.save(
  object.name,
  path.with.trailing.slash = "",
  prefix.suffix = c(prefix = "This file contains an R object called ", suffix =
    ".SavedFromR"),
  time.stamp = gsub(":", "-", Sys.time())
)
}
\description{
Save R data
}
