% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesChunks.R
\name{fetch.package}
\alias{fetch.package}
\title{Fetch an R package}
\usage{
fetch.package(
  nom_du_package,
  try.this = c("nothing", "load", "install"),
  verbose = FALSE
)
}
\description{
Fetch an R package
}
