% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{get.silly.model.chunk}
\alias{get.silly.model.chunk}
\title{Get model chunk}
\usage{
get.silly.model.chunk(
  model_flows,
  init.cond,
  init.cond.numeric.vars,
  which.flow,
  NewFrom,
  NewTo
)
}
\arguments{
\item{model_flows}{A string}

\item{init.cond}{A string}

\item{init.cond.numeric.vars}{A vector string}

\item{which.flow}{A string}

\item{NewFrom}{A string}

\item{NewTo}{A string}
}
\description{
Get model chunk
}
