% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{SaveModelInExcel}
\alias{SaveModelInExcel}
\title{Save SEIR model to Excel file}
\usage{
SaveModelInExcel(input.info.list, file_name, map.names)
}
\description{
Save SEIR model to Excel file
}
