% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.functions.R
\name{seir.n.age.classes}
\alias{seir.n.age.classes}
\title{Main box/compartment model functions}
\usage{
seir.n.age.classes(
  file.name,
  sheets.names,
  just.get.functions = FALSE,
  functions.kit = NULL,
  also.get.flows = NULL,
  agegrp.glue = "",
  CTMC.random.seeds = NULL
)
}
\arguments{
\item{file.name}{a character element.}

\item{sheets.names}{a character vector.}

\item{just.get.functions}{a logical element.}

\item{functions.kit}{a list containing one or more of the following: $differential.eqns.func, $post.processing.func, $post.processing.companion.kit, $CTMC.eqns.func, $CTMC.eqns.func.companion. This parameter is NULL by default.}

\item{also.get.flows}{a character element.}

\item{agegrp.glue}{a character element.}

\item{CTMC.random.seeds}{an integer vector.}
}
\value{
a SEIR object.
}
\description{
Main box/compartment model functions
}
