% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR.n.Age.Classes.and.friends.R
\name{SEIR.n.Age.Classes}
\alias{SEIR.n.Age.Classes}
\title{SEIR model computations}
\usage{
SEIR.n.Age.Classes(
  file.name,
  sheets.names,
  differential.eqns.func = NULL,
  post.processing.func = NULL,
  post.processing.companion.kit = NULL,
  also.get.flows = NULL
)
}
\description{
SEIR model computations
}
