% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_functions.R
\name{smooch.parms.df.into.list}
\alias{smooch.parms.df.into.list}
\title{Put model parameters into a list}
\usage{
smooch.parms.df.into.list(df.parms.1d, df.parms.2d, these.are.not.parms)
}
\arguments{
\item{df.parms.1d}{A data frame}

\item{df.parms.2d}{A data frame}

\item{these.are.not.parms}{A data frame}
}
\description{
Put model parameters into a list
}
