% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesChunks.R
\name{verbose.save}
\alias{verbose.save}
\title{Verbose save function}
\usage{
verbose.save(
  object.name,
  path.with.trailing.slash = "",
  prefix.suffix = c(prefix = "This file contains an R object called ", suffix =
    ".SavedFromR"),
  time.stamp = gsub(":", "-", Sys.time())
)
}
\arguments{
\item{object.name}{A string}

\item{path.with.trailing.slash}{A string}

\item{prefix.suffix}{A string}

\item{time.stamp}{A time object}
}
\description{
Verbose save function
}
